fix: several fixes to scripts
- 'serve' target is not available in Python Docs, so use the
    htmlview commands as alternative to check docs
- Pull translations into cpython sub-directory and then copy
    to langauge repository, as the previous patch in .tx/config
    no longer works with the new Transifex CLI tool
- Use realpath in update.sh ROOTDIR to make it easier to go back
    to the repository root directory

+++ scripts/build.sh
@@ -9,20 +9,22 @@ LANGUAGE=${1:-pt_BR}
 
 ROOTDIR="$(dirname $0)/.."
 
-cd ${ROOTDIR}
+cd "${ROOTDIR}"
 
-test -f cpython/Doc/conf.py || ( echo Unable to find proper CPython Doc folder; exit 1; )
+if ! test -f cpython/Doc/conf.py; then
+  echo Unable to find proper CPython Doc folder
+  exit 1
+fi
 
-pofiles=$(find . -maxdepth 2 -name '*.po' | sort -u)
+pofiles=$(find . -maxdepth 2 -name '*.po' | sort -u | sed -i 's|\./||')
 
 for po in ${pofiles}; do
   install -Dm644 ${po} "cpython/Doc/locales/${LANGUAGE}/LC_MESSAGES/${po}"
 done
 
-sphinx-build -b html -d build/doctrees -q --keep-going -jauto -D locale_dirs=locales -D language=pt_BR -D gettext_compact=0 -D latex_engine=xelatex -D latex_elements.inputenc= -D latex_elements.fontenc= -W cpython/Doc cpython/Doc/build/html
+sphinx-build -b html -d build/doctrees -n -a --keep-going -jauto -D locale_dirs=locales -D language=pt_BR -D gettext_compact=0 -D latex_engine=xelatex -D latex_elements.inputenc= -D latex_elements.fontenc= -W cpython/Doc cpython/Doc/build/html
 
 if [ -z "$GITHUB_ACTIONS" ]; then
-  echo "See file:/$(realpath ${ROOTDIR})/cpython/Doc/build/html/index.html"
-  echo "or serve it in http://localhost:8080 by running:"
-  echo "python3 cpython/Tools/scripts/serve.py cpython/Doc/build/html"
+  echo 'See the built documentation by running the following command:'
+  echo 'python3 -c "import os, webbrowser; webbrowser.open('file://' + os.path.realpath('cpython/Doc/build/html/index.html'))""'
 fi
